refactor: improve GET /orgConfigs endpoint unit test
Apply the same unit tests when getting data from DB and ES, so we can be sure that this endpoint works same when data comes from both of these sources.
The unit test by itself reuses most of the code now, so not need to copy/paste and adjust unit tests for DB and ES. With the current format it's much easier to write new unit tests.